Move comparator instance to where it's used
